test: split PDF visual fraction geometry coverage
